.gitlab/scripts/requirements.txt
# Python зависимости для автоматического code review с Claude
anthropic>=0.40.0
python-gitlab>=4.4.0
requests>=2.31.0
